Dratt ut i egen komponent
