+++ src/app/personside/infotabs/oppfolging/OppfolgingContainer.tsx
@@ -8,28 +8,7 @@ import { hentBaseUrls } from '../../../../redux/restReducers/baseurls';
 import { hentDetaljertOppfolging } from '../../../../redux/restReducers/oppfolging';
 import { connect } from 'react-redux';
 import PlukkRestData from '../ytelser/pleiepenger/PlukkRestData';
-import styled from 'styled-components';
-import theme from '../../../../styles/personOversiktTheme';
-import ArenaLenkerPanel from './ArenaLenkerKomponent';
-import OppfolgingDatoPanel from './OppfolgingDatoKomponent';
-import VisOppfolgingDetaljer from './OppfolgingDetaljerKomponent';
-import SykefravarsoppfolgingEkspanderbartPanel from './SykefravarsoppfolgingEkspanderbartPanel';
-import OppfolgingYtelserEkspanderbartPanel from './OppfolgingYtelserEkspanderbartPanel';
-
-const LenkeOgDatoWrapper = styled.div`
-    > *:first-child {
-        margin-bottom: ${theme.margin.layout};
-    }
-    display: flex;
-    flex-direction: column;
-    > * {
-        flex-grow: 1;
-    }
-`;
-
-interface VisningProps {
-    detaljertOppfølging: DetaljertOppfolging;
-}
+import OppfolgingVisning from './OppfolgingVisningKomponent';
 
 interface StateProps {
     baseUrlReducer: RestReducer<BaseUrlsResponse>;
@@ -46,43 +25,6 @@ interface OwnProps {
 }
 
 type Props = StateProps & DispatchProps & OwnProps;
-
-const Wrapper = styled.div`
-    display: flex;
-    flex-direction: row;
-    > *:last-child {
-        margin-left: ${theme.margin.layout};
-    }
-    > * {
-        flex-basis: 50%;
-    }
-`;
-
-const EkspanderbartPanelWrapper = styled.div`
-    display: flex;
-    flex-direction: column;
-    > * {
-        margin-top: ${theme.margin.layout};
-    }
-`;
-
-function OppfolgingVisning(props: VisningProps) {
-    return (
-        <>
-            <Wrapper>
-                <LenkeOgDatoWrapper>
-                    <ArenaLenkerPanel />
-                    <OppfolgingDatoPanel />
-                </LenkeOgDatoWrapper>
-                <VisOppfolgingDetaljer detaljertOppfølging={props.detaljertOppfølging} />
-            </Wrapper>
-            <EkspanderbartPanelWrapper>
-                <SykefravarsoppfolgingEkspanderbartPanel syfoPunkt={props.detaljertOppfølging.sykefraværsoppfølging} />
-                <OppfolgingYtelserEkspanderbartPanel ytelser={props.detaljertOppfølging.ytelser} />
-            </EkspanderbartPanelWrapper>
-        </>
-    );
-}
 
 class OppfolgingContainer extends React.PureComponent<Props> {
     componentDidMount() {
